pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "job-radar"
version = "0.1.0"
description = "Python-first job search pipeline: scan, screen, triage, apply, track."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Gene Reader" }]
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "httpx>=0.27",
  "pyyaml>=6.0",
  "markdownify>=0.13",
  "readability-lxml>=0.8",
  "lxml>=5.2",
  "anthropic>=0.39",
  "weasyprint>=62.3",
  "jinja2>=3.1",
  "python-slugify>=8.0",
  "markdown>=3.6",
]

[project.optional-dependencies]
playwright = ["playwright>=1.47"]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "mypy>=1.11",
]

[project.scripts]
job-radar = "job_radar.cli:app"
jr = "job_radar.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["job_radar"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
